Created  hook for delete opearation.
